np.int and js call function additions
* np.int was decreated. It has been replaced with np.int32
* In the javascript callback function:
        * AudioContext is resumed when the user clicks. Current policy
          in Chromiun don't let the context audio to start automacally.
        * User selection language is saved in localStorage, and restored
          the next time the page is loaded.

+++ WordMatching.py
@@ -76,7 +76,7 @@ def get_best_path_from_distance_matrix(word_distance_matrix):
             mapped_indices.append(
                 (solver.Value(estimated_words_order[word_idx])))
 
-        return np.array(mapped_indices, dtype=np.int)
+        return np.array(mapped_indices, dtype=np.int32)
     except:
         return []
